added modules for book and label class
