+++ books.json
@@ -1,1 +1,1 @@
-[{"publisher":"Publisher 1","cover_state":"bad","publish_date":"2020-09-19","archived":false,"id":585546}]
+[{"publisher":"Publisher 1","cover_state":"bad","publish_date":"2020-09-19","archived":false,"id":585546},{"publisher":"Publisher 2","cover_state":"good","publish_date":"2022-09-19","archived":false,"id":857460}]
